asr_models/openai_whisper/requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0 